Properly handle watchlist updates on page moves
Handler for TitleMoveComplete hook. This function makes it so page-moves
are handled correctly in the `watchlist` table. Prior to a MW 1.25 alpha
release when a page is moved, the new entries into the `watchlist` table
are given an notification timestamp of NULL; they should be identical to
the notification timestamps of the original title so users are notified
of changes prior to the move.