PendingReviews now sorts by importance of review
Modified files:
* WatchAnalytics.php: added configuration variables to specify which pages show
up red, orange or green in PendingReviews (egPendingReviewsRedPagesThreshold
and egPendingReviewsOrangePagesThreshold)
* includes/PendingReview.php: Change query to include sub-query to find number
of reviewers for each page. Sorts by that value.
* ext.watchanalytics.pendingreviews.css: remove green-on-hover CSS, add CSS for
color of each level of pending review criticality (red, orange, green)
* specials/SpecialPendingReviews.php: moved row-HTML generating code into
separate method, getRowHTML(), to remove duplication in getDeletedPageRow() and
getStandardChangeRow(). This method now sets a "criticality class" depending on
the number of reviewers of the page.