Remove redirects and nonexistent pages from force graph
Special:WatchAnalytics has a force-graph representation of all the
watches in the main namespace of the wiki. Prior to this commit it
included pages which didn't exist (but were in peoples' watchlist)
and redirect-pages. Neither were appropriate.

In a future release, the option of whether or not redirects should
be shown should be selectable via a form input. While selection is
not possible no-redirects is preferable.